Changed All View to use redux store, added styling, changed Single view, added Card component
